Enum ember_mug::LiquidState

pub enum LiquidState {
    Unknown,
    Empty,
    Filling,
    ColdNoTempControl,
    Cooling,
    Heating,
    TargetTemperature,
    WarmNoTempControl,
}

Represents the current state of the liquid in an Ember Mug

Variants

Unknown

The liquid state is unknown

Empty

The mug is empty

Filling

The mug is filling with liquid

ColdNoTempControl

The mug is cold and temperature control is disabled

Cooling

The mug is cooling down to the target temperature

Heating

The mug is heating up to the target temperature

TargetTemperature

The mug’s liquid is at the target temperature

WarmNoTempControl

The mug is warm and temperature control is disabled

Implementations

impl LiquidState

pub fn is_unknown(&self) -> bool

Returns true if the liquid state is Unknown.

pub fn is_empty(&self) -> bool

Returns true if the liquid state is Empty.

pub fn is_filling(&self) -> bool

Returns true if the liquid state is Filling.

pub fn is_cold_no_temp_control(&self) -> bool

Returns true if the liquid state is ColdNoTempControl.

pub fn is_cooling(&self) -> bool

Returns true if the liquid state is Cooling.

pub fn is_heating(&self) -> bool

Returns true if the liquid state is Heating.

pub fn is_target_temperature(&self) -> bool

Returns true if the liquid state is TargetTemperature.

pub fn is_warm_no_temp_control(&self) -> bool

Returns true if the liquid state is WarmNoTempControl.