Struct ember_mug::EmberMug

pub struct EmberMug { /* private fields */ }

An Ember Mug device

Implementations

impl EmberMug

pub async fn get_battery(&self) -> Result<Battery, ReadError>

Retrieves the battery percentage of the mug and other values.

impl EmberMug

pub async fn get_current_temperature(&self) -> Result<Temperature, ReadError>

Retrieves the current temperature of the mug

impl EmberMug

pub async fn get_dsk(&self) -> Result<Vec<u8>, ReadError>

Retrieves the dsk of the cup

pub async fn get_udsk(&self) -> Result<Vec<u8>, ReadError>

Retrieves the dsk of the cup

impl EmberMug

pub async fn get_liquid_level(&self) -> Result<LiquidLevel, ReadError>

Retrieves the level of liquid present in the cup

impl EmberMug

pub async fn get_liquid_state(&self) -> Result<LiquidState, ReadError>

The current state of the mug

impl EmberMug

pub async fn get_mug_color(&self) -> Result<Color, ReadError>

Retrieves the color of the mug’s LED indicator.

pub async fn set_mug_color(&self, color: &Color) -> Result<(), WriteError>

Sets the color of the mug’s LED indicator.

impl EmberMug

pub async fn get_mug_meta(&self) -> Result<MugMeta, ReadError>

Retrieves id of the mug

impl EmberMug

pub async fn get_name(&self) -> Result<String, ReadError>

Retreives the name of the mug.

pub async fn set_name(&self, name: &str) -> Result<(), WriteError>

Sets the name of the mug.

impl EmberMug

pub async fn get_ota(&self) -> Result<Ota, ReadError>

Info about the current firmware running on the mug.

impl EmberMug

pub async fn get_push_event(&self) -> Result<PushEvent, ReadError>

Events sent by the mug for the application to register to.

Call subscribe_push_events first, and prefer to use listen_push_events instead

pub async fn subscribe_push_events(&self) -> Result<(), ReadError>

Subscribe to events sent by the mug

pub async fn listen_push_events(
&self
) -> Result<impl Stream<Item = Result<PushEvent, ReadError>> + Send + '_, ReadError>

Get a stream of events sent by the mug

pub async fn unsubscribe_push_events(&self) -> Result<(), ReadError>

Unsubscribe to events sent by the mug

impl EmberMug

pub async fn get_target_temperature(&self) -> Result<Temperature, ReadError>

Retrieves the target temperature of the mug

pub async fn set_target_temperature(
&self,
temperature: &Temperature
) -> Result<(), WriteError>

Set the target temperature of the mug

impl EmberMug

pub async fn get_temperature_unit(&self) -> Result<TemperatureUnit, ReadError>

Retrieve the current unit of temperature used by the mug.

pub async fn set_temperature_unit(
&self,
temperature_unit: &TemperatureUnit
) -> Result<(), WriteError>

Set the current unit of temperature used by the mug.

impl EmberMug

pub async fn get_time_date_zone(&self) -> Result<TimeDateZone, ReadError>

Get the current date and timezone on the mug

pub async fn set_time_date_zone(
&self,
time_date_zone: &TimeDateZone
) -> Result<(), WriteError>

A sink for the mug to store the current date and timezone

impl EmberMug

pub async fn find_and_connect() -> Result<EmberMug, ConnectError>

Find and connect to the first available Ember Mug

pub async fn connect_mug(
peripheral: <Adapter as Central>::Peripheral
) -> Result<EmberMug, ConnectError>

Connect to specific Ember Mug

impl EmberMug

pub fn get_characteristic(&self, uuid: &Uuid) -> Option<&Characteristic>

Get characteristic on [EMBER_MUG_SERVICE_UUID] with given UUID

pub fn get_characteristics(&self) -> impl Iterator<Item = &Characteristic>

Get all characteristics

pub fn get_characteristic_on_service(
&self,
uuid: &Uuid,
service_uuid: &Uuid
) -> Option<&Characteristic>

Get characteristic on given service UUID with given UUID

impl EmberMug

pub async fn read_deserialize<T: BinRead + ReadEndian>(
&self,
uuid: &Uuid
) -> Result<T, ReadError>where
T::Args: Default,

Read data from given characteristic with uuid

pub async fn read(&self, uuid: &Uuid) -> Result<Vec<u8>, ReadError>

Deserialize data on given characteristic with uuid

pub async fn write<D>(
&self,
write: WriteType,
uuid: &Uuid,
data: &D
) -> Result<(), WriteError>where
D: BinWrite + WriteEndian,
<D as BinWrite>::Args: Default,

Write data to given characteristic on uuid

pub async fn command<D>(&self, uuid: &Uuid, data: &D) -> Result<(), WriteError>where
D: BinWrite + WriteEndian,
<D as BinWrite>::Args: Default,

Send command to given characteristic on uuid

pub async fn request<D>(&self, uuid: &Uuid, data: &D) -> Result<(), WriteError>where
D: BinWrite + WriteEndian,
<D as BinWrite>::Args: Default,

Send request to given characteristic on uuid